move test to separate module
